Giveaway: Fix creator.id being null in endGiveaway fn
This was due to using creator.id from dbChecks object, which is
initialized at the start of every >giveaway command. It works for other
arguments but not for "create" because there is no creator when you
run the command to create a giveaway.

Instead of using the dbChecks, we directly get the message.author.id
to resolve the issue.